extract GrantCLR to own file
