+++ app/grants/models/grant_type.py
@@ -4,7 +4,7 @@ from django.utils.translation import gettext_lazy as _
 from economy.models import SuperModel
 from grants.utils import get_upload_filename
 
-from .grant import GrantCLR
+from .grant_clr import GrantCLR
 
 
 class GrantType(SuperModel):
